+++ db/migrate/20180814142562_create_archived_file.nfs_store.rb
@@ -7,7 +7,7 @@ class CreateArchivedFile < ActiveRecord::Migration
       t.string "content_type", null: false
       t.string :archive_file, null: false
       t.string :path, null: false
-      t.integer "file_size", null: false
+      t.bigint "file_size", null: false
       t.datetime "file_updated_at"
       t.datetime "created_at", null: false
       t.datetime "updated_at", null: false
